Bug fix requested by Alex
Alex, I have fixed that bug but to make thr project work I needed to
change something in the file AdminPanel.java. Check lines 64 65 and 66
in that file. The previous commented line is your code. It was giving me
some kind of error. Look into it and see if you can fix it. At this
point the app is working but when I click the other icon on the
navigation drawer the app crashes.